+++ src/routes/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import Login from "../pages/Login";
 import Register from "../pages/RegisterUser";
 import Home from "../pages/Home";
-import DashboardAdmin from "../pages/DashboardAdmin";
+import DashboardCategory from "../pages/DashboardAdmin/Category";
 import ProtectedRoutes from "./PrivateRoute";
 import PublicRoute from "./PublicRoute";
 
@@ -13,7 +13,7 @@ function Index() {
     <BrowserRouter>
       <Routes>
         <Route element={<ProtectedRoutes />}>
-          <Route path="/dashboard" element={<DashboardAdmin />} />
+          <Route path="/dashboard" element={<DashboardCategory />} />
         </Route>
         <Route path="/" element={<Home />} />
         <Route element={<PublicRoute />}>
